+++ mcu-core/src/main/java/ru/iothub/jef/mcu/core/boards/opi/OrangePiBoardInfo.java
@@ -33,8 +33,16 @@ package ru.iothub.jef.mcu.core.boards.opi;
 
 
 enum OrangePiBoardInfo {
-    // in armbian config/boards/orange*.*
+    // https://linux-sunxi.org/GPIO
+    /**
+     * (position of letter in alphabet - 1) * 32 + pin number
+     * E.g for PH18 this would be ( 8 - 1) * 32 + 18 = 224 + 18 = 242 (since 'h' is the 8th letter).
+     */
     // PINS defenition in http://linux-sunxi.org/Category:Xunlong
+    // Boards taken from https://github.com/armbian/build/tree/master/config/boards
+    // Boards taken from https://github.com/orangepi-xunlong/orangepi-build/blob/main/external/config/boards/
+    // Pins mapping taken from https://github.com/torvalds/linux/tree/master/drivers/pinctrl/sunxi
+    // Linux doc https://www.kernel.org/doc/Documentation/pinctrl.txt
     ORANGE_PI("Orange Pi", "sun7i","Allwinner A20 dual core 1Gb SoC Wifi USB hub", null),
     ORANGE_PI_R1("Orange Pi R1","sun8i", "Allwinner H2+ quad core 256MB/512MB RAM SoC WiFi SPI 2xETH", null),
     ORANGE_PI_R1_PLUS("Orange Pi R1 Plus", "rockchip64", "Rockchip RK3328 quad core 1GB 2xGBE USB2 SPI", null),
@@ -52,13 +60,13 @@ enum OrangePiBoardInfo {
     ORANGE_PI_PC_PLUS("Orange Pi PC +", "sun8i", "Allwinner H3 quad core 1GB RAM WiFi eMMC", null),
     ORANGE_PI_PLUS("Orange Pi+", "sun8i", "Allwinner H3 quad core 1GB/2GB RAM WiFi eMMC", null),
     ORANGE_PI_PLUS_2E("Orange Pi+ 2E", "sun8i", "Allwinner H3 quad core 2GB RAM WiFi GBE eMMC", null),
-    ORANGE_PI_PRIME("Orange Pi Prime", "sun50iw2", "Allwinner H5 quad core 2GB RAM Wi-Fi/BT", null),
-    ORANGE_PI_WIN("Orange Pi Win", "sun50iw1", "Allwinner A64 quad core 1GB/2GB RAM SoC GBE WiFi/BT", null),
-    ORANGE_PI_ZERO("Orange Pi Zero", "sun8i", "Allwinner H2+ quad core 256/512MB RAM SoC WiFi SPI", OrangePiZeroBoard.class),
-    ORANGE_PI_ZERO_2("Orange Pi Zero2", "sun50iw6", "Allwinner H616 quad core 512MB/1GB RAM SoC WiFi SPI USB-C", null),
-    ORANGE_PI_ZERO_PLUS("Orange Pi Zero Plus", "sun50iw2", "Allwinner H5 quad core 512MB RAM SoC GBE WiFi SPI", OrangePiZeroBoard.class),
-    ORANGE_PI_ZERO_PLUS_2("Orange Pi Zero Plus 2", "sun8i","Allwinner H3 quad core 512MB RAM SoC Wi-Fi/BT", OrangePiZero2PlusBoard.class),
-    ORANGE_PI_ZERO_PLUS_2A("Orange Pi Zero Plus 2", "sun50iw2", "Allwinner H5 quad core 512MB RAM SoC Wi-Fi/BT", OrangePiZero2PlusBoard.class);
+    /**/ ORANGE_PI_PRIME("Orange Pi Prime", "sun50iw2", "Allwinner H5 quad core 2GB RAM Wi-Fi/BT", OrangePiPrimeBoard.class),
+    /**/ ORANGE_PI_WIN("Orange Pi Win", "sun50iw1", "Allwinner A64 quad core 1GB/2GB RAM SoC GBE WiFi/BT", OrangePiWinBoard.class),
+    /**/ ORANGE_PI_ZERO("Orange Pi Zero", "sun8i", "Allwinner H2+ quad core 256/512MB RAM SoC WiFi SPI", OrangePiZeroBoard.class),
+    /**/ ORANGE_PI_ZERO_2("Orange Pi Zero2", "sun50iw6", "Allwinner H616 quad core 512MB/1GB RAM SoC WiFi SPI USB-C", OrangePiZero2Board.class),
+    /**/ ORANGE_PI_ZERO_PLUS("Orange Pi Zero Plus", "sun50iw2", "Allwinner H5 quad core 512MB RAM SoC GBE WiFi SPI", OrangePiZeroBoard.class), // Xunlong Orange Pi Zero same pinout
+    /**/ ORANGE_PI_ZERO_PLUS_2("Orange Pi Zero Plus 2", "sun8i","Allwinner H3 quad core 512MB RAM SoC Wi-Fi/BT", OrangePiZero2PlusBoard.class),
+    /**/ ORANGE_PI_ZERO_PLUS_2A("Orange Pi Zero Plus 2", "sun50iw2", "Allwinner H5 quad core 512MB RAM SoC Wi-Fi/BT", OrangePiZero2PlusBoard.class);
 
     private final String name;
     private final String family;